feat(seo): implement canonical tags for all pages
- Added `alternates: { canonical: ... }` to `generateMetadata` in all page components.
- Refactored `src/app/[locale]/submit/page.tsx` to a Server Component to support `generateMetadata`.
- Moved client-side form logic to `src/components/SubmitToolForm.tsx`.
- Updated `params` handling to await the Promise as per Next.js 15 requirements where applicable.
