+++ src/app/[locale]/about/page.tsx
@@ -10,6 +10,9 @@ export async function generateMetadata({
   return {
     title: t('title'),
     description: t('description'),
+    alternates: {
+      canonical: `/${locale}/about`,
+    }
   };
 }
