+++ src/app/[locale]/deals/page.tsx
@@ -13,6 +13,9 @@ export async function generateMetadata({
   return {
     title: t('title'),
     description: t('description'),
+    alternates: {
+      canonical: `/${locale}/deals`,
+    }
   };
 }
